+++ packages/components/src/components/primitives/Badge/Badge.test.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
-import { screen, renderWithProviders } from 'tests/renderWithProviders';
+import { screen, renderWithProviders } from '@tests/renderWithProviders';
 import { Badge } from '.';
 
 describe('Badge', () => {
-  const givenComponentRendered = () =>
-    renderWithProviders(<Badge>Hello</Badge>);
+  const givenComponentRendered = () => renderWithProviders(<Badge>Hello</Badge>);
 
   test('should render child', () => {
     givenComponentRendered();