+++ packages/components/src/components/primitives/Link/Link.stories.tsx
@@ -1,8 +1,8 @@
-import React from 'react';
+import React, { Fragment } from 'react';
 import { Meta } from '@storybook/react';
 import { Link } from './Link';
-import { VStack, Divider } from '@chakra-ui/react';
-import { linkSizes } from './types';
+import { VStack, Divider, HStack, StackDivider } from '@chakra-ui/react';
+import { linkSizes, linkVariants } from './types';
 import * as Icons from '../Icons';
 
 export default {
@@ -14,12 +14,11 @@ const AllTemplate = () => {
   return (
     <VStack>
       {linkSizes.map((size, key) => (
-        <>
-          <Link size={size} key={key} role="link" href="#">
+        <Fragment key={key}>
+          <Link size={size} role="link" href="#">
             Size {size.toUpperCase()} link
           </Link>
           <Divider />
-
           <>
             <Link leadingIcon={Icons.IconExternalLink} size={size} role="link" href="#">
               Size {size.toUpperCase()} link with left icon
@@ -30,13 +29,31 @@ const AllTemplate = () => {
             </Link>
             <Divider />
           </>
-        </>
+        </Fragment>
       ))}
     </VStack>
   );
 };
 
 export const All = AllTemplate.bind({});
+
+const AllVariantsTemplate = () => (
+  <VStack divider={<StackDivider borderColor="gray.200" />} spacing={4}>
+    {linkVariants.map((variant, i) => (
+      <VStack key={`${variant}-${i}`}>
+        <HStack>
+          {linkSizes.map((size, i) => (
+            <Link size={size} role="link" href="#" variant={variant} key={`${size}-${i}`}>
+              {variant}
+            </Link>
+          ))}
+        </HStack>
+      </VStack>
+    ))}
+  </VStack>
+);
+
+export const AllVariants = AllVariantsTemplate.bind({});
 
 const Template = ({ showLeadingIcon, showTrailingIcon, ...args }) => (
   <VStack>
@@ -58,4 +75,5 @@ Playground.args = {
   size: 'l',
   showLeadingIcon: true,
   showTrailingIcon: true,
+  variant: undefined,
 };