+++ packages/components/src/components/primitives/Link/types.ts
@@ -1,12 +1,17 @@
 import { LinkProps as ChakraLinkProps } from '@chakra-ui/react';
 import { FC } from 'react';
+import { buttonVariants } from '../Button/types';
 import { IconProps } from '../Icons';
 
+export const linkVariants = buttonVariants;
+type LinkVariant = typeof linkVariants[number];
+
 export const linkSizes = ['s', 'm', 'l'] as const;
-type LinkSizes = typeof linkSizes[number];
+type LinkSize = typeof linkSizes[number];
 
 export interface LinkProps extends ChakraLinkProps {
-  size?: LinkSizes;
+  size?: LinkSize;
   leadingIcon?: FC<IconProps>;
   trailingIcon?: FC<IconProps>;
+  variant?: LinkVariant;
 }