+++ packages/components/src/components/primitives/Progress/Progress.tsx
@@ -1,7 +1,0 @@
-import React, { FC } from 'react';
-import { Progress as ChakraProgress } from '@chakra-ui/react';
-import { ProgressProps } from './types';
-
-export const Progress: FC<ProgressProps> = (props) => (
-  <ChakraProgress data-testid="cmpsr.progress" {...props} />
-);