+++ packages/components/src/components/primitives/Switch/Switch.test.tsx
@@ -1,16 +1,6 @@
 import React from 'react';
-import { fireEvent, screen } from '@testing-library/react';
-import { renderWithProviders } from 'tests/renderWithProviders';
+import { renderWithProviders, screen } from '@tests/renderWithProviders';
 import { Switch } from '.';
-import { useMultiStyleConfig } from '@chakra-ui/react';
-
-jest.mock('@chakra-ui/react', () => ({
-  ...(jest.requireActual('@chakra-ui/react') as any),
-  useMultiStyleConfig: jest.fn().mockReturnValue({
-    leftLabel: {},
-    rightLabel: {},
-  }),
-}));
 
 describe('Switch', () => {
   test('it should render', () => {
@@ -32,26 +22,5 @@ describe('Switch', () => {
     const labelText = screen.queryByText('Test1');
     expect(label).not.toBeFalsy();
     expect(labelText).not.toBeFalsy();
-  });
-
-  test('it should return proper styling when checked', () => {
-    const { container } = renderWithProviders(
-      <Switch labelPosition="left" label="test" />
-    );
-    const input = container.querySelector('input');
-    fireEvent.click(input);
-    expect(useMultiStyleConfig).toHaveBeenCalledWith('Switch', {
-      isChecked: true,
-    });
-  });
-
-  test('it should handle the onChange event', () => {
-    const onChange = jest.fn();
-    const { container } = renderWithProviders(
-      <Switch labelPosition="left" label="test" onChange={onChange} />
-    );
-    const input = container.querySelector('input');
-    fireEvent.click(input);
-    expect(onChange).toHaveBeenCalled();
   });
 });