+++ packages/components/src/theme/ComposerProvider/ComposerProvider.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { useTheme } from '@chakra-ui/system';
 import { ComposerProvider } from '.';
-import { theme } from 'theme';
+import { theme } from '@theme';
 
 const ThemeWrapper = () => {
   const theme = useTheme();
@@ -10,8 +10,7 @@ const ThemeWrapper = () => {
 };
 
 describe('useComposerProvider', () => {
-  const givenComponentRendered = (children: React.ReactNode) =>
-    render(<ComposerProvider>{children}</ComposerProvider>);
+  const givenComponentRendered = (children: React.ReactNode) => render(<ComposerProvider>{children}</ComposerProvider>);
 
   test('should render component and children', () => {
     givenComponentRendered(<>Children</>);